Fix messages from the mainschool page
Also add Cmd.maps to the returned transformed message from the child
update function in the main Client update function. Never really understood
what Cmd.map was for and then read some of the Elm documentation (not
the elmish docs, in some parts they assume knowledge of Elm). Anyway
Cmd is used to embed side-effects into the model update loop. For example,
sending a message back to the child can be done by embedding the message
in a Cmd. Changing the url is another thing that can be done.