Move some update logic into children
Previously some of the update logic that handled changing pages on
a button was handled in the Client.Main module. This meant
that the module had to know some of the type details of the child,
for example it needed to that the message comming from the child
was a 'Submit'  or 'Continue' then it could use Navigation to change the
url. But this breaks encapsulation. The parent shouldn't know the details
of the child message. The message should be passed to the child, and the
child then takes care of the update. So this is what we do, we pass the
message to the child and then it updates the model and returns the
thing we want to happen as a side-effect, in the Cmd parameter.
Now the Client.Main module knows nothing about the types of the
children.