Add message SetTeacherName
Add message to update the name of the teacher. Now I need to find
some way to catch the OnBlur event from the input field. I don't want to
be sending messages everytime text in the input field changes. OnBlur
apparently is the message that is sent by React when the focus is lost.