Add filesystem db
Still a WIP. We want to have a filesystem db that is basically a set of
json files that we read/write to store schools, classes and pupils.

Remove some code that was ported verbatim from the safe-bookstore
example.

Remove global.json. Don't think it's needed by .NET/code. Seems to
confuse vscode. Just use the latest SDK on the host.