Use standard formatting for records.
Previously I'd format these like I would a class or struct in C++
but this is not ideal since it ends up looking as if the { } act as
delimiters. However in F# {  } are not delimeters, and member
functions of records aren't placed within these bracse. If you
want to add a member function to a record the definition
goes outside the closing brace....this is different from C++.