Add modifed Fable fetch module
Need this to get the response body text. Fable.fetch throws an
exception when the response status is an error code and the
body text is erased and not sent to the client. This modified
function takes the text from response as is and injects the body
text into the failwith error message so the client can display it
or do what it wants with it.