Add fluentmigrator
EF was too hard to get going with....I think it was because I went
down the code first (i.e. create the migrations from the model)
road and battled with the 'dotnet ef...'
that doesn't quite work with f#. Probably picked the wrong
examples to start with. With FluentMigrator I add the up/down
transforms manually. EF promised to auto generate
those transformations from the changes to your model.