Lots of stuff
Got authorization working. Had to enable   credentials: 'include'
in the http request property so that the session cookie is used
by requests.

Renamed the pages to use hyphen instead of underscore.

Added a session type to track the jwt and user. This gets passed
down the child components on the client side so that they
can make requests back to the server with the token