Add SignOut module
Changed Login text to 'Sign In'. Seems like most web apps use 'Sign In'.

Added code to signing out of the app. Modified the view code
so that the client header at the top of the page changes the text from
there "Sign In" to "Sign Out" depending if the user is signed in or not.

Also added material design packages. Maybe we'll use this.