Make sure signout message is received
Remove the href attribute on the Sign Out header. This cause 2 messges to
be sent, which confused the processing, one was sent becuase the
url updated and the other was sent because it's in the Cmd.promise.
The one I'm interesed in is the Cmd.promise message. The url
change update message I dnon't need and ruined things.

Regardless this layout will need to change. Need to handle
the parent child composition/message flow better.