Use the Fetch api again
I don't think we need the ModifiedFableFetch for these requests
anymore. Since we are packing the error message in the response
on the server side, and not returing a 404 or some HTTP code
meaning unauthorized, but relying on the fact that we can
pass an unauthorized error into the json body of the response instead.