Rename CreateSchool module
Renamed the module since it's not really a form to create the school
anymore. It displays the properties of the tutor's school. They can edit
them on the form or create a new school in the database, but the ui flow
is the same, either for the update or create operation. Makes for a
better work flow.