omg it just worked
replaced postgres with azure sql server and i was expecting a lot of pain
but it just worked....basically substitute the ceonnection string, and change
the type of driver for Dapper. Fluent migrations still worked. Data was
written to the cloud db. Broken identity though.

Reason we did this becuase hosting azure sql server in the cloud is $5 versus
35 for postgres. per month. Wish I'd done this earlier but didn't notice that cheap
option. Everytim I clicked on the pricing tab I saw the 1000 dollar a month
option.