Got google login working
Boy what a pain. Turns out all the cors errors I was getting was due to Google
and no the the server complaining, except it wasn't clear to me where
in the flow this was happening. Why it happended is because I was using Fable.fetch
to make the get request to sign in. It makes a difference. You have to actually
navigate there in the browser ...... sigh.... On top of that working with webpack dev server in the
middle added an extra layer of confusion.