WIP: fixing database
School table had a UserId foreign key. We needed to refactor that into a
separate table called Tutor. This has fk's to the School and User table,
this mapping the tutor to a school. Anyway this spawned a lot of work.
We've had to update the database queries that used School.UserId, since
it no longer exists, and create a migration to delete that stuff from
the database, and add the new Tutor table.