Add support for samesite
Mark the cookie as SameSite.Unspecified. For more info about these
requests see https://devblogs.microsoft.com/aspnet/upcoming-samesite-cookie-changes-in-asp-net-and-asp-net-core/

It's a fairly long explanation of what the change is and took me a while
to understand. But the short version we should be checking if a browser
DOES NOT support the new samesite option (SameSite.None), and if not, then
set it as Unspecified and opt out of the SameSite flow that browsers
implement. In theory I think it is how it should. In this implementation
I ignore checking which browsers implement the old or new spec and just
opt out of SameSite processing all together.